update sockets folder structure
